lib/requirements.txt
keras
tensorflow
numpy
python-twitter